[hadolint] Exclude template files by default
By default, the hadolint runner tries to analyze files to match the `**/Dockerfile{,.*}` glob pattern.
The pattern matches `Dockerfile.erb` which is used as a template file such as in this sider/runners repository.

This change exclude `*.erb` and `*.txt` files by default.
Because such files are a template file and are not a real Dockerfile.