Migration to new Steep 0.14.0 gradually
The migration to Steep 0.14.0 is a big change because of the changes for the `.rbs` format and the syntax.
So, I think it better that we will migrate our code gradually and keep both for a while.

This change does mainly the two things:

1. Install the new version of Steep via another Gemfile, that is `Gemfile.steep`.
2. Run both type-checking (via the old and new version) on CI.

When we will complete the migration, we will be able to remove the old code.