[GolangCI-Lint] Change default linters
This change aims to make it easy to migrate from Golint to GolangCI-Lint.

Before: [stylecheck](https://github.com/dominikh/go-tools/tree/master/stylecheck)
After: [golint](https://github.com/golang/lint)

Surely, stylecheck is a replacement for golint and adds some extra rules.
But stylecheck is a 3rd-party tool, on the other hand, golint is a standard tool.
I judge that the standard tool is more suitable as a default.