[ESLint] Stop using `~/.eslintrc.*`
Summary:

- Before: Copy `sider_eslintrc.yml` to `~/.eslintrc.yml`, which is used automatically by ESLint.
- After: Specify the CLI option `--config sider_eslintrc.yml` explicitly if no config files.

It is hard to find ESLint configuration files in a project because:

- ESLint does not provide a way to find config files.
- ESLint supports different format for config files, e.g. `.eslintrc`, `.eslintrc.yml`, `package.json` and so on.
- ESLint tries to use config files in sub-directories, e.g. `sub_dir/.eslintrc.json`.

So, this change tries running the `eslint` command **two times** if no config files.
And, it parses the error message to know whether no config files, e.g. `ESLint couldn't ...`.

I think the influence for execution time is small because ESLint raises an error soon when no config files.

Fix #1073