[ESLint] Pre-install popular configs and plugins
This change aims to make Sider available with **zero configurations** of `sider.yml` and `package.json`.
It is enough for users to prepare their `.eslintrc.*` files to succeed Sider analysis.

In other words, users will be free from their dependency management for using ESLint configurations and plugins.

About this **popularity**, I decided which packages should be included in the default list by the following npm search.

- https://www.npmjs.com/search?q=eslint-config&ranking=popularity
- https://www.npmjs.com/search?q=eslint-plugin&ranking=popularity

Also, I removed some packages with low-frequency updates from the list even if the packages are popular.

Note that [eslint-plugin-vue](https://www.npmjs.com/package/eslint-plugin-vue) is still alpha requiring ESLint v7,
so I did **not** include it.
I think we should include the popular package for Vue users when it will be published with alpha.

A part of #900