[PHP_CodeSniffer] Remove deprecated `version` option
The `linter.code_sniffer.version` option has been deprecated for a while.
Also, the option was removed from the [document](https://help.sider.review/tools/php/code-sniffer).

When I look into the usage of the option, the number of repositories using it is `1` and the repository is public.
So, I believe no problems if we remove it.