Replace :pull_number with :refspecs
This change allows analyses for any Git repositories with additional
refspecs. The previous implementation assumes that the Git repositories
are on GitHub, so it had prepared the `:pull_number` parameter to
construct an additional refspec. We should support other platforms, such
as GitLab. So, Runners will receive the new parameter `:refspecs`. From
this change, the Runners users should construct refspecs by themselves
and pass them to Runners.