Consider line range in `Runners::Changes#include?`
Some tools have a line range (start and end) for an issue.
But the current implementation of `Runners::Changes#include?` checks only a start line.
In the following case, an issue is never reported.

Code:

```ruby
def foo     # L1
  p "Hi"    # L2 <--- changed line, but never reported!
end         # L3
```

Issue: "Too long method with L1-3"

This change aims to fix the bug by checking a line range of an issue.