[RuboCop] Set up default config when no user config
When a user specify a custom config file via `sider.yml`,

```yaml
linter:
  rubocop:
    config: my-rubocop.yml
```

the default config currently is copied to `.rubocop.yml` in the user's repository on staring analysis:
<https://github.com/sider/runners/blob/9239a21e2c38c69d3b03231fbe7201cb48c1bc14/images/rubocop/default_rubocop.yml>

But this copied config is ignored (because `rubocop --config my-rubocop.yml` is executed)
and `meowcop` is installed unnecesarily.
This change aims to fix that needless behavior.

Note that no smoke test cases are added because the following case satisfies them already:
<https://github.com/sider/runners/blob/9239a21e2c38c69d3b03231fbe7201cb48c1bc14/test/smokes/rubocop/with_safe_cops/sideci.yml#L5>

Fix #2018