Loosen unsatisfied constraints for npm packages
This change aims to loosen the behavior when npm-installed packages are unsatisfied with our version constraints.

Before: an analysis error is raised
After: fall back to the default (pre-installed) version with a warning

I think the change is reasonable because raising an error is too strict.

Also, this change does not use the `npm ls` command because:
- it will be changed with npm v7 (see also #1963)
- it is hard to control

Instead of `npm ls`, this checks `node_modules/{package_name}/` directories to get installed versions.